requirements.txt
tqdm
redis
psycopg2-binary
mwparserfromhell
nltk
numpy